add availble Zones in finger print

+++ Dawem.Domain/Entities/Core/Zone.cs
@@ -15,7 +15,7 @@ namespace Dawem.Domain.Entities.Core
         public string Name { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public decimal? Radius { get; set; }
+        public double? Radius { get; set; }
 
     }
 }
