+++ Dawem.Models/Dtos/Core/Zones/UpdateZoneDTO.cs
@@ -7,6 +7,6 @@
         public bool IsActive { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public decimal? Radius { get; set; }
+        public double? Radius { get; set; }
     }
 }
