+++ Dawem.Models/Response/Core/Zones/GetZoneByIdResponseDTO.cs
@@ -7,7 +7,7 @@
         public bool IsActive { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public decimal? Radius { get; set; }
+        public double? Radius { get; set; }
